Adjusted Dracula Theme
I adjusted the colors a bit as i thought some looked off. In my opinion some of them looked off so i adjusted them to where i think they look better together. For example the white looked a little to dark which made it seem like a murky gray and the black was darker than its bright equivalent also most of the colors are the same which looked weird in my setup. Any feedback is appreciated this is just some edits i thought looked better.