Fix tmux support for -p.
Previously we indiscriminately sent escaped request sequences to tmux.
This fails in the event that multiple terminals are attached to the same
session since it interleaves responses from multiple terminals. This
patch attempts to discover the controlling terminal and send the request
sequences directly to it. The responses are subsequently collected from
the tmux pseudo terminal.