Fix color palette for PaperColor themes
Ensure that the palette follows the original, upstream definition[0],
as the previous defined ones were quite a bit off. While the colors
used are in the PaperColor palette, their place in the palette
definition was wrong/scrambled.

I compared the now fixed ones with the VIM variant, which the
original PaperColor project defines itself, by simply opening a plain
text file there and using `cat` on the same file in another shell,
quickly switching between them to ensure both are rendered the same.

Additionally, I compared the adapted palette here with the PaperColor
theme of the foot editor, looks good now.

0: https://github.com/NLKNguyen/papercolor-theme#-colors

Signed-off-by: Thomas Lamprecht <thomas@lamprecht.org>