Only source theme.sh in interactive fish sessions
I had the same issue as
https://www.reddit.com/r/neovim/comments/xr65ea/awk_error_when_calling_terminal_commands/,
which is caused by theme.sh trying to apply the theme when vim is
executing shell commands.

This commit amends the suggested default theme.sh loading logic for the
fish shell to only apply when the shell is interactive. This seems like
a generally sensible thing for a terminal theme manager regardless.